move existing models for stimmzettel to own directory
