+++ wls-gui-wahllokalsystem/src/components/dse/BaseStimmzettelZusammenfassungCard.vue
@@ -45,7 +45,7 @@ import type { Wahlvorschlag } from "@/types/wahlvorschlaege/Wahlvorschlag.ts";
 
 import { computed } from "vue";
 
-import { StimmzettelGueltigkeitEnum } from "@/types/dse/StimmzettelGueltigkeitEnum.ts";
+import { StimmzettelGueltigkeitEnum } from "@/types/dse/persistedStimmzettel/StimmzettelGueltigkeitEnum.ts";
 
 const props = defineProps<{
   listenstimmen: Wahlvorschlag[];
