+++ wls-gui-wahllokalsystem/src/composables/dse/stimmzettelMapper.ts
@@ -5,11 +5,11 @@ import type {
   StimmzettelOfTeamDTO,
   WahlvorschlagDTO,
 } from "@/api/wls-clients/generated-ergebnismeldung-api";
-import type { Beschlussfassung } from "@/types/dse/Beschlussfassung.ts";
-import type { Beschlussgrund } from "@/types/dse/Beschlussgrund.ts";
-import type { Kandidat } from "@/types/dse/Kandidat.ts";
-import type { Stimmzettel } from "@/types/dse/Stimmzettel.ts";
-import type { Wahlvorschlag } from "@/types/dse/Wahlvorschlag.ts";
+import type { Beschlussfassung } from "@/types/dse/persistedStimmzettel/Beschlussfassung.ts";
+import type { Beschlussgrund } from "@/types/dse/persistedStimmzettel/Beschlussgrund.ts";
+import type { Kandidat } from "@/types/dse/persistedStimmzettel/Kandidat.ts";
+import type { Stimmzettel } from "@/types/dse/persistedStimmzettel/Stimmzettel.ts";
+import type { Wahlvorschlag } from "@/types/dse/persistedStimmzettel/Wahlvorschlag.ts";
 
 export function useStimmzettelMapper() {
   function toModel(dto: StimmzettelOfTeamDTO): Stimmzettel {
