+++ wls-gui-wahllokalsystem/src/composables/dse/stimmzettelService.ts
@@ -1,4 +1,4 @@
-import type { Stimmzettel } from "@/types/dse/Stimmzettel.ts";
+import type { Stimmzettel } from "@/types/dse/persistedStimmzettel/Stimmzettel.ts";
 
 import {
   Configuration,
