--- wls-gui-wahllokalsystem/src/types/dse/Stimmzettel.ts
@@ -1,15 +1,0 @@
-import type { Beschlussfassung } from "@/types/dse/Beschlussfassung.ts";
-import type { Beschlussgrund } from "@/types/dse/Beschlussgrund.ts";
-import type { Wahlvorschlag } from "@/types/dse/Wahlvorschlag.ts";
-
-import { StimmzettelGueltigkeitEnum } from "@/types/dse/StimmzettelGueltigkeitEnum.ts";
-
-export interface Stimmzettel {
-  stimmzettelkennung: number;
-  wahlvorschlaege: Wahlvorschlag[];
-
-  invalideVotes: number;
-  gueltigkeit: StimmzettelGueltigkeitEnum;
-  beschlussvorschlag: Beschlussgrund[];
-  beschlussfassung: Beschlussfassung | null;
-}
