+++ wls-gui-wahllokalsystem/src/types/dse/persistedStimmzettel/Beschlussfassung.ts
