+++ wls-gui-wahllokalsystem/src/types/dse/persistedStimmzettel/Beschlussgrund.ts
