+++ wls-gui-wahllokalsystem/src/types/dse/persistedStimmzettel/Kandidat.ts
