+++ wls-gui-wahllokalsystem/src/types/dse/persistedStimmzettel/StimmzettelGueltigkeitEnum.ts
